+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 - Removed `unsafe` requirement from `Memory::allocate` method
+- Renamed `Hazard` struct to `Pointers`
 
 ## [0.2.0] - 2017-03-24